Replace a non-fixed width lookbehind with an atomic group
Fixes #1. One of the regex patterns contains a positive lookbehind with a quantifier ('*'), making its width non-fixed. Many regex engines will complain about that, as it isn't a widely supported feature.
We can replace the positive lookbehind with an atomic group, achieving the same behavior: asserting the presence of the group before the capture, without capturing it too.